+++ src/components/charts/BarChartMulti.js
@@ -20,6 +20,7 @@ import { getBarSize } from '@src/helper/chartUtility'
 
 import XAxis from '@Components/charts/cartesian/XAxis'
 import YAxis from '@Components/charts/cartesian/YAxis'
+import CartesianGrid from '@Components/charts/cartesian/CartesianGrid'
 
 const Box = styled.article`
   &:not(:last-child) {
@@ -34,6 +35,7 @@ const BarChartMulti = ({
   yDataKey,
   theme,
   isPercent,
+  textMap,
 }) => {
   const newYDataKey = [].concat(yDataKey)
   const colors = getColorsByTheme(theme, newYDataKey.length)
@@ -66,7 +68,7 @@ const BarChartMulti = ({
     <div>
       <Heading size="18" style={{ marginBottom: '30px' }}>{title}</Heading>
 
-      <cdmCommon.LegendList data={legendData} />
+      <cdmCommon.LegendList data={legendData} textMap={textMap} />
       {
         isEmpty(data)
           ? (
@@ -88,14 +90,12 @@ const BarChartMulti = ({
                       top: 0, right: 0, bottom: 0, left: (i === 0 ? 0 : YAxisWidth * -1),
                     }}
                   >
-                    <Rechart.CartesianGrid
-                      vertical={false}
-                      stroke={colorV1.$grey04}
-                    />
+                    <CartesianGrid vertical={false} />
                     <XAxis dataKey={xDataKey} />
                     <YAxis tickFormatter={tickFormatter} width={YAxisWidth} domain={domain} />
                     <Rechart.Tooltip
                       isPercent={isPercent}
+                      textMap={textMap}
                       content={TooltipBox}
                       cursor={{ fill: ColorSet['Chart-Hover']['grey-hover'] }}
                     />
@@ -129,6 +129,7 @@ BarChartMulti.defaultProps = {
   yDataKey: [],
   theme: Themes.ThemeArrangePrimarySea,
   isPercent: false,
+  textMap: {},
 }
 
 BarChartMulti.propTypes = {
@@ -137,6 +138,7 @@ BarChartMulti.propTypes = {
   xDataKey: PropTypes.string,
   yDataKey: PropTypes.oneOfType([PropTypes.string, PropTypes.arrayOf(PropTypes.string)]),
   isPercent: PropTypes.bool,
+  textMap: PropTypes.shape({}),
   theme: PropTypes.oneOf([
     'blue', 'green', 'compare',
     Themes.ThemeArrangePrimarySea, Themes.ThemeArrangeSecondaryTeal,