+++ src/components/charts/RadarChart.js
@@ -1,171 +1,123 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types'
-import Highcharts from 'highcharts';
-import HCMore from 'highcharts/highcharts-more'
-import HighchartsReact from 'highcharts-react-official';
-import _ from 'lodash'
+import React from 'react'
+import * as Rechart from 'recharts'
 import { colorV1 } from '@src/assets/styles/variables'
-import { errorMessage, getTextStyleForHighcharts } from '@src/helper/chartUtility'
-
+import * as commonTag from '@Components/common/cdmCommon'
 import {
-  getColorsByTheme,
-  Themes,
+  getColorsByTheme, Themes,
 } from '@Components/ChartColor'
+import TooltipBox from '@Components/tooltip/TooltipBox'
 
-if (typeof Highcharts === 'object') {
-  HCMore(Highcharts)
+import _ from 'lodash'
+import PropTypes from 'prop-types'
+
+import PolarAngleAxis from '@Components/charts/cartesian/PolarAngleAxis'
+
+const tooltipContent = ({
+  active, payload, nameKey, textMap,
+}) => {
+  if (active) {
+    return (
+      <TooltipBox
+        payload={payload.reverse()}
+        nameKey={nameKey}
+        textMap={textMap}
+      />
+    )
+  }
+  return null
 }
 
-class RadarChart extends Component {
-  constructor(props) {
-    super(props)
+tooltipContent.defaultProps = {
+  active: false,
+  payload: {},
+  textMap: {},
+  nameKey: 'name',
+}
 
-    const {
-      title, width, height,
-      radarCategory, groupData, patientData, legendOpen,
-      theme,
-    } = this.props
+tooltipContent.propTypes = {
+  active: PropTypes.bool,
+  payload: PropTypes.arrayOf(PropTypes.shape({})),
+  textMap: PropTypes.shape({}),
+  nameKey: PropTypes.string,
+}
 
-    this.colors = getColorsByTheme(theme)
-
-    this.options = {
-      chart: {
-        polar: true,
-        type: 'area',
-        width,
-        height,
-        marginLeft: 100,
-        marginTop: 50,
-      },
-      title: {
-        text: title,
-      },
-      xAxis: {
-        categories: radarCategory,
-        tickmarkPlacement: 'on',
-        lineWidth: 0,
-        labels: {
-          distance: 14,
-          style: getTextStyleForHighcharts(colorV1.$grey08),
-        },
-        gridLineWidth: 1,
-        gridLineColor: colorV1.$grey06,
-      },
-      yAxis: {
-        gridLineInterpolation: 'polygon',
-        gridLineWidth: 1,
-        gridLineColor: colorV1.$grey06,
-        lineWidth: 1,
-        min: 0,
-        labels: {
-          enabled: false,
-        },
-        tickAmount: 3,
-        lineColor: colorV1.$grey06,
-      },
-      series: [
+// Legend 클릭 시 Radar 를 toggle 하려면 Radar의 props 인 hide 을 사용하세요
+const RadarChart = ({
+  data,
+  nameKey,
+  dataKey,
+  theme,
+  textMap,
+}) => {
+  const newDataKey = [].concat(dataKey).reverse()
+  const colors = getColorsByTheme(theme, newDataKey.length).reverse()
+  const legendData = _.chain(newDataKey)
+    .map((entry, index) => ({ color: colors[index], text: entry }))
+    .value().reverse()
+  return (
+    <div>
+      <commonTag.LegendList data={legendData} textMap={textMap} />
+      <Rechart.RadarChart
+        data={data}
+        width={700}
+        height={520}
+        margin={{}}
+        cx={350}
+        cy={270}
+        outerRadius={250}
+        style={{
+          margin: '0 auto',
+        }}
+      >
+        <Rechart.Tooltip
+          content={tooltipContent}
+        />
+        <Rechart.PolarGrid stroke={colorV1.$grey04} />
+        <PolarAngleAxis dataKey={nameKey} />
         {
-          name: 'Group',
-          color: this.colors[1],
-          data: groupData,
-          pointPlacement: 'on',
-        },
-        {
-          name: 'Patient',
-          color: this.colors[0],
-          data: patientData,
-          pointPlacement: 'on',
-        },
-      ],
-      legend: {
-        enabled: legendOpen,
-        align: 'left',
-        verticalAlign: 'top',
-        layout: 'horizontal',
-        symbolHeight: 10,
-        symbolWidth: 10,
-        symbolRadius: 5,
-        itemStyle: getTextStyleForHighcharts(colorV1.$grey08),
-        reversed: true,
-        x: -18.5,
-        itemDistance: 24,
-      },
-      tooltip: {
-        enabled: false,
-      },
-      plotOptions: {
-        series: {
-          marker: {
-            enabled: true,
-            symbol: 'circle',
-            radius: 5,
-          },
-          fillOpacity: 0.3,
-        },
-      },
-    }
-  }
-
-  renderRadarChart = (options) => (
-    <HighchartsReact highcharts={Highcharts} options={options} />
+          newDataKey.map((entry, index) => (
+            <Rechart.Radar
+              key={`radar${entry}`}
+              dataKey={entry}
+              stroke={colors[index]}
+              fill={colors[index]}
+              strokeWidth={3}
+              fillOpacity={0.3}
+              dot
+            />
+          ))
+        }
+      </Rechart.RadarChart>
+    </div>
   )
-
-  checkDataValidation = () => {
-    const { groupData, patientData } = this.props
-    if (_.isEmpty(groupData) || _.isEmpty(patientData)) return 'haveData'
-    if (!Array.isArray(groupData) || !Array.isArray(patientData)) return 'typeOfVariable'
-    return null
-  }
-
-
-  render() {
-    if (this.checkDataValidation()) {
-      return <div>{errorMessage(this.checkDataValidation())}</div>
-    }
-
-    return <div>{this.renderRadarChart(this.options)}</div>
-  }
 }
-
-// title, width, height,
-// radarCategory, groupData, patientData, legendOpen,
 
 RadarChart.defaultProps = {
-  title: null,
-  width: 1200,
-  height: 1200,
-  radarCategory: [
-    'visit_info',
-    'visit_history',
-    'lab',
-    'echo',
-    'drug',
-    'spect',
-    'demo',
-    'comorbidity',
-    'cabgpci',
-    'vitalsign',
-  ],
-  groupData: [],
-  patientData: [],
-  legendOpen: true,
+  data: [],
+  nameKey: 'name',
+  dataKey: ['value', []],
   theme: Themes.ThemeComparePrimarySea2,
+  textMap: {},
+  margin: {},
 }
 
 RadarChart.propTypes = {
-  title: PropTypes.string,
-  width: PropTypes.number,
-  height: PropTypes.number,
-  radarCategory: PropTypes.arrayOf(PropTypes.string),
-  groupData: PropTypes.arrayOf(PropTypes.number),
-  patientData: PropTypes.arrayOf(PropTypes.number),
-  legendOpen: PropTypes.bool,
+  data: PropTypes.arrayOf(PropTypes.shape({})),
+  nameKey: PropTypes.string,
+  dataKey: PropTypes.oneOfType([PropTypes.string, PropTypes.arrayOf(PropTypes.string)]),
   theme: PropTypes.oneOf([
     Themes.ThemeComparePrimarySea, Themes.ThemeComparePrimarySea1,
     Themes.ThemeComparePrimarySea2, Themes.ThemeComparePrimarySea3,
     Themes.ThemeCompareSecondaryTeal, Themes.ThemeCompareSecondaryTeal1,
     Themes.ThemeCompareSecondaryTea2, Themes.ThemeCompareSecondaryTeal3,
   ]),
+  textMap: PropTypes.shape({}),
+  margin: PropTypes.shape({
+    top: PropTypes.number,
+    right: PropTypes.number,
+    bottom: PropTypes.number,
+    left: PropTypes.number,
+  }),
 }
 
-export default RadarChart;
+export default RadarChart